New (X)HTML Syntax for Wiki Macros
see https://community.openproject.com/projects/openproject/work_packages/24511
