+++ lib/open_project/text_formatting/transformers/image_attachment_transformer.rb
@@ -1,0 +1,73 @@
+#-- encoding: UTF-8
+#-- copyright
+# OpenProject is a project management system.
+# Copyright (C) 2012-2015 the OpenProject Foundation (OPF)
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License version 3.
+#
+# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
+# Copyright (C) 2006-2013 Jean-Philippe Lang
+# Copyright (C) 2010-2013 the ChiliProject Team
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+#
+# See doc/COPYRIGHT.rdoc for more details.
+#++
+
+module OpenProject
+  module TextFormatting
+    module Transformers
+      class ImageAttachmentTransformer < TextTransformer
+        def post_process(fragment, options)
+          obj = options[:object]
+          attachments = options[:attachments]
+          if attachments.nil? && obj.respond_to?(:attachments)
+            attachments = obj.attachments
+          end
+          if !attachments.nil? && !attachments.empty?
+            fragment.xpath('///img').each do |node|
+              process_img node, attachments
+            end
+          end
+          fragment
+        end
+
+        private
+
+        def process_img(node, attachments)
+          attributes = node.attributes
+          # TODO:basename of filename
+          filename = attributes[:src].downcase
+          if found = attachments.detect { |att| att.filename.downcase == filename }
+            description = found.description
+            if !description.blank?
+              update_attributes node, description, attributes
+            end
+          end
+        end
+
+        def update_attributes(node, description, attributes)
+          if attributes[:title].nil?
+            node.set_attribute(:title, description)
+          end
+          if attributes[:alt].nil?
+            node.set_attribute(:alt, description)
+          end
+        end
+      end
+    end
+  end
+end
