+++ lib/open_project/text_formatting/transformers/redmine_wiki_transformer.rb
@@ -1,0 +1,52 @@
+#-- encoding: UTF-8
+#-- copyright
+# OpenProject is a project management system.
+# Copyright (C) 2012-2015 the OpenProject Foundation (OPF)
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License version 3.
+#
+# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
+# Copyright (C) 2006-2013 Jean-Philippe Lang
+# Copyright (C) 2010-2013 the ChiliProject Team
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+#
+# See doc/COPYRIGHT.rdoc for more details.
+#++
+
+module OpenProject
+  module TextFormatting
+    module Transformers
+      class RedmineWikiTransformer < TextTransformer
+        def process(fragment, options)
+          result = Nokogiri::XML.fragment ''
+          fragment.children.each do |node|
+            if node.text?
+              text = Redmine::WikiFormatting.to_html(options[:format], node.to_s,
+                                                     object: options[:object],
+                                                     attribute: options[:attr],
+                                                     edit: options[:edit])
+              result.add_child Nokogiri::XML.fragment text
+            else
+              result.add_child node
+            end
+          end
+          return result
+        end
+      end
+    end
+  end
+end
